feat: add ability to create a new memory lane on homepage
